requirements.txt
cv==1.0.0
cycler==0.10.0
demjson==2.2.4
Django==2.2.7
django-ckeditor==5.9.0
django-crispy-forms==1.10.0
django-csvexport==1.1.5
django-filter==2.4.0
django-js-asset==1.2.2
django-queryset-csv==1.1.0
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
nose==1.3.7
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
sklearn-pandas==2.0.4
sqlparse==0.4.1
threadpoolctl==2.1.0
unicodecsv==0.14.1
xlrd==2.0.1
xmltodict==0.12.0
